ut for inner key off 0 (#1673)

+++ components/kvengine/src/util.rs
@@ -314,7 +314,6 @@ pub(crate) mod test_util {
     }
 
     impl KeyBuilder {
-        // prefix: reserved to generate TiDB keys (prefix as "t_").
         pub fn new(keyspace_id: u32, enable_inner_key_off: bool, prefix: &str) -> Self {
             Self {
                 keyspace_id,
@@ -323,7 +322,7 @@ pub(crate) mod test_util {
             }
         }
 
-        pub fn i_to_inner_key(&self, i: i32) -> OwnedInnerKey {
+        pub fn i_to_inner_key(&self, i: usize) -> OwnedInnerKey {
             let v = if self.enable_inner_key_off {
                 self.i_to_key(i)
             } else {
@@ -332,14 +331,14 @@ pub(crate) mod test_util {
             OwnedInnerKey::new(Bytes::from(v))
         }
 
-        pub fn i_to_outer_key(&self, i: i32) -> Vec<u8> {
+        pub fn i_to_outer_key(&self, i: usize) -> Vec<u8> {
             let mut key = ApiV2::get_txn_keyspace_prefix(self.keyspace_id);
             key.extend_from_slice(&self.i_to_key(i));
             key
         }
 
         #[inline]
-        pub fn i_to_key(&self, i: i32) -> Vec<u8> {
+        pub fn i_to_key(&self, i: usize) -> Vec<u8> {
             format!("{}key{:06}", self.prefix, i).into_bytes()
         }
     }
